requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.10.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.10
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
